+++ Bugsnag/Assets/Bugsnag/Editor/BugsnagAddScriptingSymbol.cs
@@ -1,9 +1,8 @@
 using UnityEditor;
-using UnityEngine;
 namespace BugsnagUnity.Editor
 {
     [InitializeOnLoad]
-    public class BugsnagAddScriptingSymbol : MonoBehaviour
+    public class BugsnagAddScriptingSymbol
     {
         private const string DEFINE_SYMBOL = "BUGSNAG_UNITY_WEB_REQUEST";