+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 
-## 8.4.0 (2025-12-10)
+## TBD
+
+You can now automatically upload all your Android and iOS files for stacktrace symbolication in a few simple steps. Please see our online docs for information on how to enable this feature or script it yourself: https://docs.bugsnag.com/platforms/unity/showing-full-stacktraces/
+
+### Enhancements
+
+- Add auto symbol uploads for Android and Xcode builds. [#871](https://github.com/bugsnag/bugsnag-unity/pull/871)
+
+## 8.4.0 (2025-12-1)
 
 ### Enhancements